move test config
